requirements.txt
pillow
pytesseract
opencv-python
PyQt6
python-docx
openpyxl